ASoC: msm: q6dspv2: use correct variable type to store ION buff size"
The size of the physical memory allocated for ION buffers
are of type size_t. Change updates the type of variables
sent to ION drivers to size_t to avoid any mismatch.

+++ sound/soc/msm/qdsp6v2/q6voice.h
@@ -95,7 +95,7 @@ enum {
 struct mem_buffer {
 	dma_addr_t		phys;
 	void			*data;
-	uint32_t		size; /* size of buffer */
+	size_t			size; /* size of buffer */
 };
 
 struct share_mem_buf {
